Do not zoom if BoxZoomTool selects range <= 5 pixels
As far as I know, there is no way to cancel out of a BoxZoomTool
selection after the mouse is held down (and then there's no way
to undo the zoom after the mouse is released). To get around this,
use the behavior of the zoom on other plotting tools like matplotlib:
if either dimension of the range selection is <= 5 pixels, do not
apply the zoom.